Move some compilation-related types to their own folder.
