+++ Expressions/Literal.cs
@@ -30,6 +30,7 @@
 
 using System;
 using System.Collections.Generic;
+using MetaphysicsIndustries.Solus.Compiler;
 
 namespace MetaphysicsIndustries.Solus.Expressions
 {
